Events API update: add event + get event

+++ src/events/models.py
@@ -13,7 +13,7 @@ EventsTbl = sa.Table(
     sa.Column('name', sa_types.String(length=70), nullable=False),
     sa.Column('date', sa_types.TIMESTAMP),
     sa.Column('place', sa_types.String(length=100)),
-    sa.Column('price', sa_types.NUMERIC(4, 2), default=0.0),
+    sa.Column('price', sa_types.FLOAT, default=0.0),
     sa.Column('max_players', sa_types.INT, default=1000),
     sa.Column('created_at', sa_types.TIMESTAMP, default=datetime.utcnow),
     sa.UniqueConstraint('name', 'place', 'date', name='uq__event__name__place__date')
